requirements.txt
requests>=2.28.0
phonenumbers>=8.13.14
pyyaml>=6.0